pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "openai-agent-sdk-supervisor"
version = "0.1.0"
description = "A multi-agent AI supervisor built with the OpenAI Agents SDK"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Your Name", email = "your.email@example.com" },
]
keywords = [
    "openai",
    "ai",
    "agents",
    "supervisor",
    "multi-agent",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "openai-agents>=0.2.8",
    "openai>=1.60.0",
    "tavily-python>=0.5.0",
    "python-dotenv>=1.1.1",
    "braintrust<0.6.0",
    "rich>=14.2.0",
    "autoevals<0.1.0",
    "modal>=1.2.0",
    "fastapi[standard]>=0.119.1",
    "ipython>=9.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "ruff>=0.3.0",
]

[project.scripts]
openai-agent-sdk-supervisor = "src.local_runner:main"

[project.urls]
Homepage = "https://github.com/yourusername/openai-agent-sdk-supervisor"
Repository = "https://github.com/yourusername/openai-agent-sdk-supervisor"
Issues = "https://github.com/yourusername/openai-agent-sdk-supervisor/issues"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.black]
line-length = 88
target-version = ["py310", "py311", "py312"]
include = '\.pyi?$'

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
]
ignore = [
    "E501", # line too long (handled by black)
    "B008", # do not perform function calls in argument defaults
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
